Restructure project to client and server
