+++ .gitignore
@@ -8,4 +8,4 @@ public
 .vscode
 
 # jest coverage reports
-coverage 
+coverage
