+++ client/.dockerignore
@@ -15,3 +15,6 @@ node_modules
 jest.config.js
 coverage
 tests
+
+# Compiled
+public
