+++ client/constants/PATHS.js
@@ -1,0 +1,5 @@
+const path = require('path');
+
+module.exports = {
+  ROOT: path.resolve(__dirname, '..'),
+};
