+++ client/jest.config.js
@@ -1,5 +1,4 @@
 const path = require('path');
-const { ROOT } = require('./constants/PATHS');
 
 module.exports = {
   clearMocks: true, // Automatically clear mock calls and instances before every test.
@@ -11,6 +10,6 @@ module.exports = {
   testEnvironment: 'node',
   setupFilesAfterEnv: [
     './node_modules/jest-enzyme/lib/index.js', // Setup jest-enzyme library
-    path.resolve(ROOT, 'tests', 'setupTests.js'), // Configure enzyme react adapter
+    path.resolve(__dirname, 'tests', 'setupTests.js'), // Configure enzyme react adapter
   ],
 };
