+++ client/tests/Calendars.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import moment from 'moment';
 
-import { Calendars } from '../client/src/components/Calendars/Calendars';
+import { Calendars } from '../src/components/Calendars/Calendars';
 
 describe('Calendars component', () => {
   const calendars = shallow(
