+++ package.json
@@ -5,12 +5,6 @@
   "main": "index.js",
   "scripts": {
     "test": "jest",
-    "dev-server": "env NODE_ENV=development nodemon index.js",
-    "start-server": "node index.js",
-    "build-dev": "webpack --config config/webpack.config.dev.js",
-    "build-prod": "webpack --config config/webpack.config.prod.js",
-    "start-dev": "webpack-dev-server --config config/webpack.config.dev.js",
-    "seed": "node database/seed.js",
     "eslint": "eslint './**/*.js?(x)'"
   },
   "repository": {
@@ -24,40 +18,11 @@
   },
   "homepage": "https://github.com/Air-Tree-n-Tree/Reservation-Service#readme",
   "devDependencies": {
-    "@babel/core": "^7.8.7",
-    "@babel/preset-env": "^7.8.7",
-    "@babel/preset-react": "^7.8.3",
-    "babel-jest": "^25.1.0",
-    "babel-loader": "^8.0.6",
-    "css-loader": "^3.4.2",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.1.0",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^2.5.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^25.1.0",
-    "jest-enzyme": "^7.1.2",
-    "style-loader": "^1.1.3",
-    "webpack": "^4.42.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
-  },
-  "dependencies": {
-    "axios": "^0.19.2",
-    "express": "^4.17.1",
-    "html-webpack-plugin": "^3.2.0",
-    "moment": "^2.24.0",
-    "mongoose": "^5.9.4",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0",
-    "react-redux": "^7.2.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "reselect": "^4.0.0"
+    "eslint-plugin-react-hooks": "^2.5.0"
   }
 }
