+++ server/index.js
@@ -1,4 +1,4 @@
-const app = require('./server/server');
+const app = require('./src/server');
 
 app.listen(3002, () => {
   console.log('Server listening on: 3002');
