+++ server/src/database/utils/generateAvailability.js
